feat(adminaccess): make admin-dashboard page only accessible to admin
Made a new page admin-dashboard which is serverside and uses server action auth. Made admincontent clientside where only admins have access. Moved logout functionality here. Made /admin still show login google auth. Now ready for showing adminthings to admins
