+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-	"conventionalCommits.scopes": ["frontend", "format", "setup", "songform", "colortheme", "pwabanner"]
+	"conventionalCommits.scopes": ["frontend", "format", "setup", "songform", "colortheme", "pwabanner", "adminaccess"]
 }
